remove aztec-node from aztec.js, move contract tree back to aztec-rpc
