+++ yarn-project/aztec.js/src/contract/contract_tree/index.ts
@@ -1,1 +1,2 @@
 export * from './contract_tree.js';
+export * from './function_tree_data.js';
